Add a device type, which can be looked up by ID.
The intention is that this will do all the actual message-parsing. It
should also save its state at regular intervals and load state when
instantiated.

I'm allowing for multiple devices because it's easy to do, despite the
fact that I suspect I'm the only person who still has a working Plantcube
and I've only got one of them.