Process updates, applying them to Device properties. Retain update timestamps.
The intention here is that the methods for sending .../accepted updates can
scan through all properties for the ones with the given update timestamp and
send those as accepted updates.

No follow-up processing is done on the updates (e.g. sending recipes when
seeing recipe ID 1). This should be done after accepting updates.