Implement save/restore of Device. Remove Device.userOffset.
Save/restore: I played with various options, but just exporting all the fields
(except the channel and the MQTT client, which we don't want to save) turned
out to be least bad. A custom marshaler is included for valueWithTimestamp, to
make it empty when the update timestamp is zero.

User offset isn't actually used by the device (the string is not included
anywhere in the ESP8266 firmware). As far as I can tell: the _app_ sends user
offset. The AWS-side Agrilution code calculates total offset from this,
updates the shadow and this update is sent as a delta to the device. When the
offset is changed, the sequence of messages seen from AWS is:

1. update/accepted, with user_offset but no total offset and a clientToken
   that's not that of the Plantcube (presumably that of the app).
2. update/accepted, 4s later, with total_offset and no user_offset and no
   clientToken. This is presumably the Agrilution code updating the shadow.
3. update/delta with total_offset.