Subscribe explicitly to topics that start with a dollar sign.
Mosquitto (and maybe/probably other brokers) won't deliver topics that start
with a dollar sign as a result of a wildcard ("#") subscription. From testing,
a $foo/+/bar/baz subscription is fine. We therefore explicitly subscribe to
our two $aws/... topics and to the wildcard topic. Since the wildcard matches
the dollar-sign topics absolutely fine on our end and since we don't want the
message handler to be called multiple times, we only set a message handler for
the wildcard subscription.