Recipe creation and marshalling with tests.
The recipes created by this don't exactly match what the Agrilution code
produces. In all observed Agrilution recipes, one layer has two blocks and the
other has one block. The two-block layer begins with a single-period block
of length 24h and then has a two-period block with lengths 15.5h/8.5h. The
single-block layer has two periods in its block, with lengths 15.5h/8.5h.
As far as I can tell, it should be perfectly reasonable to just provide a
single block with two periods for both layers - and I'll probably try that
later. Nevertheless, for now, we're generating both layers with the same
values and that includes having a "skip" block with a single whole-day period
and a normal day/night two-period block.

The Plantcube recipe-parsing code parses this fine and produces sensible
offsets. The recipe also won't be too long. It's limited to 3000 bytes on the
Plantcube side and we're at 105.