Send replies to agl/prod/.../recipe/get with the current recipe.
agl/prod/.../recipe is the only message that's not in JSON format - it's the
raw binary recipe. I've implemented this by making a msgReplyBinary extension
of msgReply which has a Marshal() method. At reply-sending time, we then check
if it's a msgReplyBinary and marshal it with the method instead of via
encoding/json. This is a bit messy. Alternatives considered:

1. Everything gets a Marshal method, which is just repeated boilerplate for
   the other replies.
2. Publish this one reply directly rather than passing it back up the caller
   chain.
3. Have recipe implement MarshalJSON but sneakily not return JSON. Unclear if
   this would actually work.

All of these are varying degrees of ick, as is the current solution. I've not
thought of a solution that is ick-free. (1) is potentially less ick depending
on your valuation of boilerplate vs of reflection.