Two new actions, push and pop, for swapping message order.
Particularly for deltas, AWS sometimes sends .../update/accepted, then
.../update/delta and sometimes sends the same messages in the opposite order.
There appears to be no logic in which order is selected. As far as I can tell
from the Mongoose and Agrilution code, the ordering will make no difference.
(Indeed, for most examples seen so far, one of them has a clientToken that's
not the Plantcube's token, so it'll be ignored.)