Generate a new recipe 2 mins after a planting or harvesting event.
On closer examination of the dumps, it looks like this was at least happening
in the June dumps. It's not always _exactly_ 2 minutes, but it's within a few
seconds. I seem to remember the same thing wasn't happening in later dumps,
but that might be my memory or a consequence of Agrilution's shutdown.