Stop passing time around quite so much purely for testing. Use a mock instead.
We still pass time around in message processing, where it seems helpful for
avoiding inconsistent timestamps in messages, but we're now mocking time for
tests. This in turn should also allow us to use a timer for recipe queuing
without the immense pain it seems like that would otherwise incur.