Implement manualAction for setting AWS version.
We already had a method to bump it, but in the case of gaps in our packet
dumps, we need to be able to set it, because otherwise that's a long series of
bumps.  The bumps are still useful, don't want to worry about which AWS
version is active any time we ignore an accepted message for an app update.