Cancel the watering timer if watering happens organically.
We're doing this by cancelling it whenever we see the valve be something other
than closed. This leaves a race condition (we could send the RPC just as the
device is about to trigger watering). But it's a small race and the other
option (working through current recipe when scheduling watering to determine
if we need it or not) sounds painful, so here we are.