Allow one second of slop when comparing JSON messages.
Our messages have consistent timestamps within a message. AWS's messages don't
and additionally, may be at a different second to the one we expect them in
(because in real life, network latency is non-zero). We therefore find a
timestamp in our message and replace any timestamps that are 1 less or 1 more
than our timestamp with our timestamp before doing our JSON comparison.