Deal with MQTT publish messages split across multiple TCP packets.
This is hacky. When this happens in our dumps, we see a series of 1024-byte
packets, then a <1024 byte packet. What we _should_ do is probably to switch
to piping the packets into a channel, have a single reader for all the packets
in the channel and they'll coalesce themselves. That was more work than I
wanted to do. This works with our dumps and they're the only thing we need to
read.