Calculate nutrient wanted.
This calculates using a PID controller, tuned in the nutrient spreadsheet,
to aim for an EC value of 1450. As the EC values vary with temperature and our
smoothing correction is not completely accurate, we additionally smooth the
values before passing to the PID controller.

Show the raw and smoothed values as well as the nutrient wanted in the status
page.